Modularize `Plot` code
This allows the code to be shared across platforms.
