+++ core/shared/src/main/scala/chartreuse/NumberFormat.scala
@@ -16,11 +16,7 @@
 
 package chartreuse
 
-import scala.scalajs.js
-import scala.scalajs.js.annotation.*
-
-@js.native
-@JSGlobal("Intl.NumberFormat")
-class NumberFormat extends js.Object {
-  def format(value: Double): String = js.native
+/** Interface that will be implemented by platform specific number formatting */
+trait NumberFormat {
+  def format(value: Double): String
 }
